refactor: split bins routes into modular directory structure
